Standardize embeds rich text tests
